FEATURE: Implement Keyboard event recognition
Sets up basic framework for application to recognize keystrokes.

The `Button` component is responsible for rendering buttons on the screen.
The `Keyboard` component is responsible for instantiating all `keys` and
managing the state of the `keys`.

These `keys` lives as an array of `Key` objects.

Each `key` in the `keys` array should have a

* `code` - corresponding to the `KeyboardEvent.key` [interface](https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/key)
* `position` - corresponding to the `KeyboardEvent.location` [interface](https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/location)
* `size` - the size of the keyboard, based on a unit value of `u`
* `name` - the string to display on the `Button` component

Further, each key starts with the `pressed` state set to `false`. This
`pressed` state is used to communicate to the `Button` component whether or
not to display this key as pressed or not.